feat: 目录整理,删除无关的文件
