+++ src/config/index.ts
@@ -2,3 +2,16 @@ export const serverConfig = {
     port: 3000,
     hostname: 'localhost'
 }
+
+// // 判断环境
+// if (process.env.NODE_ENV && process.env.NODE_ENV !== 'development') {
+//     config = proConfig
+// }
+
+export const dbConfig = {
+    host: 'sugarat.top',
+    port: 3306,
+    user: 'sugar',
+    password: 'a123456',
+    database: 'auto_course'
+}
