+++ src/db/modal/index.ts
@@ -1,0 +1,5 @@
+export interface User {
+    id: number
+    account: string
+    password: string
+}
