+++ src/lib/middleware/initServer.ts
@@ -1,5 +1,5 @@
-import { matchRouter } from './../router'
-import { SuperHttpRequest, SuperHttpResponse } from '@/types'
+import { matchRouter } from './../../router'
+import { SuperHttpRequest, SuperHttpResponse } from '@/lib/types'
 
 export default function (req: SuperHttpRequest, res: SuperHttpResponse) {
     const route = matchRouter(req)
