--- test/index.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Document</title>
-    <script src="https://cdn.bootcss.com/jquery/2.2.4/jquery.min.js"></script>
-</head>
-
-<script>
-    $.ajax({
-        url: 'http://localhost:3000/api/test',
-        method: 'POST',
-        success: function(res) {
-            console.log(res);
-        }
-    })
-</script>
-
-<body>
-
-</body>
-
-</html>
