--- test/qsTest.js
@@ -1,3 +1,0 @@
-const qs = require('query-string')
-
-console.log(qs.parse('https://www.baidu.com/s?wd=abc&rsv_spt=1&rsv_iqid=0xa4c51a42000c606a&issp=1&f=8&rsv_bp=1&rsv_idx=2&ie=utf-8&tn=baiduhome_pg&rsv_enter=1&rsv_dl=tb&rsv_sug3=4&rsv_sug1=1&rsv_sug7=100&rsv_sug2=0&rsv_btype=i&prefixsug=abc&rsp=6&inputT=1132&rsv_sug4=1644'));
